remove duplicated component
